mudando tipo do estoque da cerveja, implementando entrada e saida de cerveja no estoque

+++ backend/src/main/java/br/com/chadschoperia/model/BeerModel.java
@@ -34,6 +34,6 @@ public class BeerModel implements Serializable {
 	private Double valuePerMug;
 
 	@Column(name = "stock", nullable = false)
-	private Long stock;
+	private Double stock;
 
 }
